% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushbar.R
\name{pushbar-buttons}
\alias{pushbar-buttons}
\alias{pushbar_open}
\alias{pushbar_close}
\title{Pushbar Buttons}
\usage{
pushbar_open(session, id = c("left", "right", "top", "bottom"))

pushbar_close(session)
}
\arguments{
\item{session}{A valid Shiny session.}

\item{id}{Id of pushbar to open.}
}
\description{
Open and close pushbar programatically.
}
